redis_reference.py renamed as
redis_callback_data.py. Docstrings added. RedisError is handled.
